requirements.txt
# ============================================================
# LE GÉNÉRATEUR DIGITAL — BACKEND REQUIREMENTS
# Backend FastAPI + SQLModel + PostgreSQL + IA + Automations
# ============================================================

# --- Core Framework ---
fastapi==0.115.0
uvicorn==0.30.6

# --- Database & ORM ---
sqlmodel==0.0.22
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9

# --- Prisma ORM ---
prisma==0.15.0

# --- Auth & Security ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# --- Validation ---
email-validator==2.1.1
pydantic==2.11.10
typing-extensions==4.12.2
pydantic-settings>=2.0.0

# --- Environment ---
python-dotenv==1.0.1

# --- API HTTP ---
requests==2.32.3
httpx==0.27.0
openai>=1.0.0

# --- Logging & Utils ---
rich==13.9.4
loguru==0.7.2

# --- PDF / Documents ---
reportlab==4.2.2
python-docx==1.1.2

# --- Node env (required by Prisma) ---
nodeenv==1.9.1
